Update log statements and move 'DataSourceMetadataHelper' to dataSourceMetadata directory
